Uglify terser copy (#256)
Summary:
<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

**Summary**

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->

Add support for using [terser](https://github.com/fabiosantoscode/terser/). `terser` is a fork of `uglify-es`, which is not really maintained; and maintains API compatibility with `uglify-es`.

In order to avoid duplicating configuration like default options in `metro-minify-uglify`, this PR adds a parameter to `metro-minify-uglify` that allows passing in an alternative minifier module, and the new package `metro-minify-terser` calls this, passing in `terser`.

This is based on discussion in https://github.com/facebook/metro/issues/230.

After earlier submitting https://github.com/facebook/metro/pull/243, I created this alternative PR to address comments from rafeca

**Test plan**

<!-- Demonstrate the code is solid. Example: The exact commands you ran and their output, screenshots / videos if the pull request changes UI. -->

The tests are simply copied from `metro-minify-uglify`, since the API is and the results should be identical.
Pull Request resolved: https://github.com/facebook/metro/pull/256

Reviewed By: pvdz

Differential Revision: D9908604

Pulled By: rafeca

fbshipit-source-id: 30dfe4f8cc6b614e50874be51aad0883a1997061