Add custom serializer (#309)
Summary:
<!-- Thanks for submitting a pull request! Please provide enough information so that others can review your pull request. The two fields below are mandatory. -->

**Summary**
We would like to bundle an entryfile differently than it is currently done in Metro, specifically in development. Without a pluggable serializer, we wouldn't be able to perform bundle splitting in development, which is impactful to our development speed.

This PR alleviates bundle sizing problem by introducing a customSerializer instead of using the `plainJSBundle`.

Some minor question, do I have to implement this in the `build` public method in the server?

<!-- Explain the **motivation** for making this change. What existing problem does the pull request solve? -->

**Test plan**
- Wrote a `customSerializer` and test it against the new change in metro. Our bundle was able to do what we needed it to accomplish.
Pull Request resolved: https://github.com/facebook/metro/pull/309

Reviewed By: rafeca

Differential Revision: D12981378

Pulled By: mjesun

fbshipit-source-id: 77ac4bd8bcc9d9338e0582390947627e9bbc8b17