Pass caller option to babel when transforming. Fixes #190 (#319)
Summary:
**Summary**

Passing this option allows us to conditionally return a different babel config according to the platform. For example:

```js
// babel.config.js
module.exports = api => {
  const platform = api.caller(caller => caller.platform);

  if (platform === 'android') {
    // return some config
  }

  // return some config
};
```

This is especially useful when using the new JSC on Android where we don't need to run all of the transforms.

Fixes #190

**Test plan**

Tested in a freshly inited React native app. The `caller` is correctly set to `{ caller: 'metro', platform: 'ios' }`
Pull Request resolved: https://github.com/facebook/metro/pull/319

Differential Revision: D13237123

Pulled By: rafeca

fbshipit-source-id: 950c6a10a59ba8c8fd5b894f02b1a4bf9d724723