Bring back terser minifier (#510)
Summary:
**Summary**

This brings back `metro-minify-terser` which was removed because it wasn't used. Now to make sure it is used we'll make it the default.

`terser` is a fork of `uglify-es` which is not maintained. Updating this is required to remove es6 transforms since `uglify-es` currently crashes on in my app when not transforming es6. It would probably be possible to track down the bug but at this point it's probably easier and better long term to migrate to `terser`. It is now the default minifier for webpack and seem pretty stable.

**Test plan**

Tested in my app to make sure minifying works when disabling es6 transforms.
Pull Request resolved: https://github.com/facebook/metro/pull/510

Differential Revision: D19578475

Pulled By: cpojer

fbshipit-source-id: 07bead308153198604c5846896e8eedd2417422a