Remove AssetResolutionCache and platform specific asset
Summary:
This diff removes support for platform specific assets like "image.ios.png" as we are not using this feature at Facebook.

Further, this diff removes the AssetResolutionCache module and a bunch of complex friends. Instead of building a dynamic list of possible resolutions by crawling the file system, we now have a config that lists common asset resolution sizes and then resolves them in O(1) by looking the files up in `hasteFS`.

In follow-ups I will remove further code and eventually get rid of "node-haste", and hopefully also start converging on ModuleGraph and other data structures to reduce the duplication between Metro and metro-buck.

Reviewed By: motiz88

Differential Revision: D19970453

fbshipit-source-id: e3164514a1845fe8c17252561910a86922ccdecb