Blackhole `jest-environment-jsdom` and `@jest/types`
Summary:
These two packages are 7 MiB in total (jsdom is huge) yet we are not using them. We are using `jest-environment-node` and we aren't using TypeScript. Jest always shipped with jsdom by default. When I turned Jest into a monorepo, I created the node environment but decided to keep jsdom as a default (this makes sense: most people use Jest to test web code). Both are automatically shipped with the `jest` package regardless of use. In parallel to this change, I'm discussing with the Jest maintainers to make the `node` environment the default in a future major release, as well as removing `jsdom` from the default install.

This will also make it easier to upgrade Jest.

Reduces node_modules from 227 MiB to 220 MiB.

Reviewed By: motiz88

Differential Revision: D20814953

fbshipit-source-id: 8cbde3073c755f5449ecf94a5880269c129f6c50