Fail gracefully if a code frame for an internal stack cannot be retrieved
Summary: When developing Metro it is possible to cause an error that is thrown in a node.js builtin module. When we try to print a nice error message with a code frame, the `fs` call will fail and therefore mask the real error, leading to a very confusing experience for the developer. This diff fixes it.

Reviewed By: motiz88

Differential Revision: D21044016

fbshipit-source-id: 57df565a5f4c4b4db4eb0286a18987126f767298