Add file length as a header for each bytecode module
Summary:
This diff adjusts the bytecode format to encode the length of the bytecode buffer as a UInt32 in front of each bytecode unit. This makes it so that the clients do not need to know anything about Hermes bytecode and how to read data from its header. It's using a UInt32 because that takes 4 bytes and therefore won't cause alignment issues.

NOTE: The bytecode format will change in the future, I'm just making minimal changes from now on.

Reviewed By: motiz88

Differential Revision: D21233233

fbshipit-source-id: 690958cfb151afca4b73c82b4b5174f39db1ad5a