Codemod `require.requireActual` to `jest.requireActual`
Summary:
This method was deprecated in 2017 because it cannot be flow-typed and in favor of `jest.requireActual`. This diff codemods the remaining callers and removes support for tracking it as part of the `dependencyExtractor`.

This is necessary to upgrade to Jest 26: D21369069

I'm working with the Jest team to hold off on more breaking changes like these until Jest 27.

Reviewed By: yungsters

Differential Revision: D21373371

fbshipit-source-id: 64bc6dd5419ef72329ffeeba21f7821d0c72d707