Expose splitCondition
Summary:
The goal of this stack is to unify the metro-core `collectDependencies` with an internal fork.

The internal fork of the `collectDependencies` implementation exposes the `splitCondition` of conditionally included JSResources.

This diff adds the `splitCondition` to the interface of the `collectDependencies` implementation. The implementation of `metro-core` keeps the current behavior of not initializing the `splitCondition`.

Reviewed By: motiz88

Differential Revision: D23540048

fbshipit-source-id: 1baafe926c2930df64a87f8a548943fb12a70df3