Improve FileStore performance on Windows
Summary:
The main thread stores the transformed files in the `FileStore` which writes the file to the disk.

Profiling the Fb4aBundle on Windows showed that the main thread spends a lot of time in `fs.closeSync`, waiting for NTFS to write the file. My theory is that NTFS isn't as good at handling small files as other file systems. Writing the file synchronous is problematic because it blocks the main-thread that as a result becomes unresponsive to accept responses from workers or schedule new work.

Changing the `FileStore` implementation to async removes this bottleneck on Windows and didn't show any regressions on my On-Demand servers (but no improvements either).

Reviewed By: cpojer

Differential Revision: D25167921

fbshipit-source-id: 86614dd4dc939c404208d0ee1c1125fd8e14bf92