Fix update throttling of TerminalReporter
Summary:
The `throttle` of the terminal reporter only throttled the internal update of the `progress` object but not the message printed to the terminal. The result was that CPU profiles showed very frequent sync flushes to the console which block the main thread and create unnecessary load on the main thread (and potential process switches on the CPU).

This fix ensures that the whole `update` is throttled and not only of the internal state.

This further fixes that the build command uses the unthrottled event that then is internally throttled by the Terminal Reporter

Reviewed By: cpojer

Differential Revision: D25169585

fbshipit-source-id: e561ed0322aa6edffc039c27783eb21483740125