Format log messages using printf format
Summary:
In browsers and Node.js, console methods accept a printf format string as a first argument:

```
console.log('Hello %s', 'world');
//> Hello world
```

This is not currently supported in the terminal reporter in Metro, which just outputs the log messages as they come from the client (as an array of values, most often strings).

This adds support for it in Metro by using the `util.format` function from Node.js (which is the same method used by the console under the hood).

Reviewed By: motiz88

Differential Revision: D27154865

fbshipit-source-id: 9e7e2e9d07fd4e273f25396e00fb944d488c4580