Add hermesParser option in metro and transformer
Summary:
Allow the caller to use the HermesParser (compiled to WASM)
in its `babel: true` mode instead of the Babel parser directly.

This is configurable via the `hermesParser` option to the transformer,
and it is off by default.

Sometimes we have to parse TS, in which case we use Babel regardless.
This is acceptable because we really want Hermes parser for parsing
newer Flow syntax and for performance.
TS is a minority of the code we're parsing.

Reviewed By: motiz88

Differential Revision: D27367573

fbshipit-source-id: 732f0d386c6816caab5400a3725964f9669c543d