Extract JavaScript transformation in transform-worker
Summary:
This diff is part of a stack that re-structures the metro-transform-worker with the goal to make the transformation pipeline(s) more visible as well as what metadata is added by each step. The ultimate goal is to model the pipeline as a set of transformation-steps where the output of one step is the input of the next step (e.g. the asset step transforms an asset to a JS artifact that than is transformed by the JS step).

This diff extracts the JavaScript transformation logic in the metro-transform-worker into its own function.

Reviewed By: motiz88

Differential Revision: D28384721

fbshipit-source-id: 00e02a7d94ce05c4164381649dd736744614f561