Refactor: Rename Device._getPageId to Device._mapToDevicePageId
Summary:
## Rationale
The name _getPageId is a bit ambiguous. When it really does is translate a reloadable page id to a device page id. So, this diff renames the method to more closely match its responsibilities.

## Metro Inspector Proxy Context
Metro Inspector Proxy is a middleware that sits on the Metro server. It connects the Hermes debugger running on Flipper with the Hermes JS VM running on a device. Metro Inspector Proxy intercepts/transforms/forwards the messages coming from the device to Flipper, and vice versa.

In Metro Inspector Proxy, [we expose a virtual/dummy reloadable page (with id = -1)](https://fburl.com/code/c086aa1k). What is a page? AFAIK, it's [another term for a JS VM running on device](https://fburl.com/code/5ptfp3n8).

This dummy page is [what the Hermes Debugger on Flipper connects to](https://fburl.com/code/84osfyln). Then, Metro Inspector Proxy intercepts all incoming messages from the device, and replaces the dummy page id with [the real page id (i.e: the id of the last page that was connected to the proxy)](https://fburl.com/code/twgdshj1) via the _getPageId method.

Reviewed By: p-sun

Differential Revision: D29236899

fbshipit-source-id: 3a2e55529aae17b9c630afc3125fb9f3598acc7b